requirements.txt
# torch (Base Image)
# torchvision (Base Image)
botorch>=0.10.0
gpytorch>=1.11.0
rtamt>=0.3.0
onnx>=1.15.0
onnxruntime>=1.17.0
pymoo>=0.6.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
pyyaml>=6.0
tqdm>=4.66.0
scipy>=1.11.0
Levenshtein>=0.23.0
wandb
huggingface_hub
python-dotenv
git+https://github.com/Verified-Intelligence/onnx2pytorch@fe7281b9b6c8c28f61e72b8f3b0e3181067c7399
onnxoptimizer>=0.3
onnxsim>=0.4.31
skl2onnx>=1.14
termcolor>=2.3.0
sortedcontainers>=2.4
psutil>=5.9.5
gurobipy>=10
